e/mail-relay/deliver/requirements.txt
dnspython==1.12.0
Flask==0.10.1
gevent==1.0.2
redis==2.10.3
supervisor==3.1.3
urllib3==1.10.4
